Update user handler.

+++ internal/pkg/models/user.go
@@ -1,10 +1,10 @@
 package models
 
 type User struct {
-	UID       int64
+	UID       int
 	Username  string `json:"username"`
 	Email     string `json:"email",valid:"email"`
-	Rating    int64  `json:"rating"`
+	Rating    int    `json:"rating"`
 	AvatarUrl string `json:"avatar_url"`
 	Password  string `json:",omitempty",valid:"length(8|20)"`
 }
